@@ -12,7 +12,7 @@
     "description": "A coroutine framework that focuses on hyperspeed and flexible, specifically use for build microservices and middlewares.",
     "license": "MIT",
     "require": {
-        "php": ">=7.3",
+        "php": ">=8.0",
         "ext-json": "*",
         "ext-openssl": "*",
         "ext-pdo": "*",
@@ -48,7 +48,7 @@
         "hyperf/ide-helper": "2.2.*",
         "hyperf/testing": "2.2.*",
         "mockery/mockery": "^1.0",
-        "phpstan/phpstan": "^1.0.1",
+        "phpstan/phpstan": "^1.0",
         "swoole/ide-helper": "dev-master",
         "symfony/var-dumper": "^5.1"
     },
