@@ -20,20 +20,8 @@ use Throwable;
 
 class RPCExceptionHandler extends ExceptionHandler
 {
-    /**
-     * @var StdoutLoggerInterface
-     */
-    protected $logger;
-
-    /**
-     * @var FormatterInterface
-     */
-    protected $formatter;
-
-    public function __construct(StdoutLoggerInterface $logger, FormatterInterface $formatter)
+    public function __construct(protected StdoutLoggerInterface $logger, protected FormatterInterface $formatter)
     {
-        $this->logger = $logger;
-        $this->formatter = $formatter;
     }
 
     public function handle(Throwable $throwable, ResponseInterface $response)
